Split up slam_march2023.yaml into 4 different yaml files
slam_march2023.yaml too way to long to run on Biowulf in one go. Thus,
we split up the file by the age variable (age and relative age) and the
subset (all mice and sex/strain subset). I think that in the future, it
is better to limit each dataset in the yaml file to one age variable. It
makes reading and creating the yaml files much easier.