Appears function were renamed and combined cox was made.
There were uncommmited changes after a long while, thus I am not too
sure all the changes that have been made since the previous commit or if
the code works. It looks a combined cox survival model is made across
outcomes. There also appears to be a renaming of a few functions.